EvoNN-Shared/pyproject.toml
[project]
name = "evonn-shared"
version = "0.1.0"
description = "Shared substrate contracts and helpers for the EvoNN monorepo"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evonn_shared"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
